+++ src/ga_auth_handlers.cpp
@@ -1379,10 +1379,25 @@ namespace sdk {
 
     auth_handler::state_type get_unspent_outputs_for_private_key_call::call_impl()
     {
-        if (m_net_params.is_liquid() || m_net_params.is_electrum()) {
-            throw user_error("Sweeping is not yet implemented for Liquid or singlesig wallets");
+        if (m_net_params.is_liquid()) {
+            throw user_error("Sweeping is not yet implemented for Liquid wallets");
         }
-        m_result = m_session->get_unspent_outputs_for_private_key(m_details);
+        auto private_key = json_get_value(m_details, "private_key");
+        auto password = json_get_value(m_details, "password");
+
+        std::vector<unsigned char> private_key_bytes;
+        bool compressed;
+        try {
+            std::tie(private_key_bytes, compressed)
+                = to_private_key_bytes(private_key, password, m_net_params.is_main_net());
+        } catch (const std::exception&) {
+            throw user_error(res::id_invalid_private_key);
+        }
+        auto public_key_bytes = ec_public_key_from_private_key(gsl::make_span(private_key_bytes));
+        if (!compressed) {
+            public_key_bytes = ec_public_key_decompress(public_key_bytes);
+        }
+        m_result = m_session->get_unspent_outputs_for_private_key(private_key_bytes, public_key_bytes, compressed);
         return state_type::done;
     }